refactor: rename connector auth provider registry
